cleaned up files in perception
