+++ src/perception/CMakeLists.txt
@@ -160,7 +160,7 @@ include_directories(
 
 ## Mark executable scripts (Python etc.) for installation
 ## in contrast to setup.py, you can choose the destination
-catkin_install_python(PROGRAMS src/perception_node.py src/occupancy_map_simple.py src/practice.py
+catkin_install_python(PROGRAMS src/perception_node_simple.py src/perception_node.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
